+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - the `sea_orm` crate feature allows to use `MagnetLink`, `TorrentFile` and `TorrentID`
   in sea_orm models
 - `TorrentFile`, `DecodedTorrent` and `DecodedInfo` implement `PartialEq`
+- `MagnetFile` implements `FromStr`
 
 ## Version 0.4.0 (2025-11-10)