+++ CHANGELOG.md
@@ -22,6 +22,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `MagnetFile` sorts and deduplicates the trackers provided in the magnet URL;
   `MagnetFile::trackers` returns them in an order, which is not specified
 
+### Fixed
+
+- `MagnetLink` equality is no longer affected by the exact representation of the
+  magnet URL; if the name, hash, and (sorted/deduped) trackers are equal, then
+  two magnets are considered equal
+
 ## Version 0.4.0 (2025-11-10)
 
 This release is focused on stricter parsing of torrents and magnets, and